PlantDialog created
Viewing a plant from the availability page used to open it up in a dialog. For mobile screens, this caused a lot of wasted space. Now, the data pops up in a full screen dialog
